separate config class files
